See the modulesexample folder.
The hnnengine is the starting point for the project.
This dynamically loads all the kernels in the directory 'kernels'
A valid kernel file is specified by inserting a new file whose name is
the name of the kernel and it defines exactly one function by the same name.
After the parsing the dictonary called kernels will have the keys as the
module names and the the value is the function object to that kernel